requirements.txt
crewai
streamlit
tavily-python
python-dotenv
litellm==1.61.20
